Separate address from netmask or prefix length before validation
The previous commits will help you start a pot that has been
configured with alias networking and addresses with netmasks or prefix
lengths, but stopping a pot may fail when it tries to remove addresses
from network interfaces.  If the netmask or prefix length is present,
ifconfig(8) won't like it very much when you try to `-alias`
a.k.a. `delete`.  So we parse the recorded addresses yet again in
order to ensure proper cleanup of network addresses.

Fixes #268